harasser/requirements.txt
twilio
google-cloud-bigquery
pandas